removed chain from integration test directory structure
