+++ packages/periphery/package.json
@@ -16,13 +16,12 @@
     "verify": "hardhat etherscan-verify --sleep --solc-input",
     "gasReport": "REPORT_GAS=true OPTIMIZER_ENABLED=true yarn test:integration",
     "test": "HARDHAT_SHOW_STACK_TRACES=true hardhat test test/unit/**/*",
-    "test:integration": "FORK_ENABLED=true FORK_BLOCK_NUMBER=17433155 hardhat test test/integration/mainnet/**/*",
-    "test:integrationArbitrum": "FORK_ENABLED=true FORK_NETWORK=arbitrum FORK_BLOCK_NUMBER=243648015 hardhat test test/integration/l2/**/*",
-    "test:integrationArbitrumMI": "FORK_ENABLED=true FORK_NETWORK=arbitrum FORK_BLOCK_NUMBER=243648015 hardhat test test/integration/mainnet/**/*",
-    "test:integrationBase": "FORK_ENABLED=true FORK_NETWORK=base FORK_BLOCK_NUMBER=21067741 hardhat test test/integration/l2/**/*",
+    "test:integration": "FORK_ENABLED=true FORK_BLOCK_NUMBER=17433155 hardhat test test/integration/**/*",
+    "test:integrationArbitrum": "FORK_ENABLED=true FORK_NETWORK=arbitrum FORK_BLOCK_NUMBER=243648015 hardhat test test/integration/**/*",
+    "test:integrationBase": "FORK_ENABLED=true FORK_NETWORK=base FORK_BLOCK_NUMBER=21067741 hardhat test test/integration/**/*",
     "coverage": "hardhat coverage --testfiles 'test/unit/**/*'",
-    "coverage:integration": "FORK_ENABLED=true FORK_BLOCK_NUMBER=17433155 hardhat coverage --testfiles 'test/integration/mainnet/**/*'",
-    "coverage:integrationArbitrum": "FORK_ENABLED=true FORK_NETWORK=arbitrum FORK_BLOCK_NUMBER=243648015 hardhat coverage --testfiles 'test/integration/l2/**/*'",
+    "coverage:integration": "FORK_ENABLED=true FORK_BLOCK_NUMBER=17433155 hardhat coverage --testfiles 'test/integration/**/*'",
+    "coverage:integrationArbitrum": "FORK_ENABLED=true FORK_NETWORK=arbitrum FORK_BLOCK_NUMBER=243648015 hardhat coverage --testfiles 'test/integration/**/*'",
     "lint": "eslint --fix --ext '.ts,.js' ./ && solhint 'contracts/**/*.sol' --fix",
     "format": "prettier -w .",
     "clean": "rm -rf cache artifacts types/generated deployments/localhost",
