+++ packages/periphery/test/integration/MultiInvoker/Arbitrum.test.ts
@@ -1,10 +1,8 @@
 import { ethers } from 'hardhat'
 import { BigNumber, constants, utils } from 'ethers'
-import { Address } from 'hardhat-deploy/dist/types'
 import { loadFixture } from '@nomicfoundation/hardhat-network-helpers'
-import HRE from 'hardhat'
 
-import { IERC20Metadata__factory, IKeeperOracle } from '../../../../types/generated'
+import { IERC20Metadata__factory, IKeeperOracle } from '../../../types/generated'
 
 import { RunInvokerTests } from './Invoke.test'
 import { RunOrderTests } from './Orders.test'
@@ -18,15 +16,15 @@ import {
   fundWalletUSDC,
   PYTH_ADDRESS,
   USDC_ADDRESS,
-} from '../../../helpers/arbitrumHelpers'
-import { deployPythOracleFactory } from '../../../helpers/setupHelpers'
-import { parse6decimal } from '../../../../../common/testutil/types'
+} from '../../helpers/arbitrumHelpers'
+import { deployPythOracleFactory } from '../../helpers/setupHelpers'
+import { parse6decimal } from '../../../../common/testutil/types'
 import {
   advanceToPrice as advanceToPriceImpl,
   createPythOracle,
   PYTH_ETH_USD_PRICE_FEED,
-} from '../../../helpers/oracleHelpers'
-import { time } from '../../../../../common/testutil'
+} from '../../helpers/oracleHelpers'
+import { time } from '../../../../common/testutil'
 import { KeeperOracle, PythFactory } from '@perennial/v2-oracle/types/generated'
 
 const ORACLE_STARTING_TIMESTAMP = BigNumber.from(1684116265)
