+++ packages/periphery/test/integration/MultiInvoker/Mainnet.test.ts
@@ -1,18 +1,17 @@
 import { ethers } from 'hardhat'
 import { BigNumber, constants, utils } from 'ethers'
-import { Address } from 'hardhat-deploy/dist/types'
 import { loadFixture } from '@nomicfoundation/hardhat-network-helpers'
-import { parse6decimal } from '../../../../../common/testutil/types'
+import { parse6decimal } from '../../../../common/testutil/types'
 
 import { ChainlinkContext } from '@perennial/v2-core/test/integration/helpers/chainlinkHelpers'
 import { OracleVersionStruct } from '@perennial/v2-oracle/types/generated/contracts/Oracle'
 import { CHAINLINK_CUSTOM_CURRENCIES } from '@perennial/v2-oracle/util/constants'
 
-import { IERC20Metadata__factory, IOracle__factory, KeeperOracle, PythFactory } from '../../../../types/generated'
+import { IERC20Metadata__factory, IOracle__factory, KeeperOracle, PythFactory } from '../../../types/generated'
 import { RunInvokerTests } from './Invoke.test'
 import { RunOrderTests } from './Orders.test'
 import { RunPythOracleTests } from './Pyth.test'
-import { createInvoker, deployProtocol, InstanceVars, resetBtcSubOracle, resetEthSubOracle } from './setupHelpers'
+import { createInvoker, deployProtocol, InstanceVars } from './setupHelpers'
 import {
   CHAINLINK_ETH_USD_FEED,
   DSU_ADDRESS,
@@ -22,9 +21,9 @@ import {
   fundWalletUSDC,
   PYTH_ADDRESS,
   USDC_ADDRESS,
-} from '../../../helpers/mainnetHelpers'
-import { createPythOracle, PYTH_ETH_USD_PRICE_FEED } from '../../../helpers/oracleHelpers'
-import { deployPythOracleFactory } from '../../../helpers/setupHelpers'
+} from '../../helpers/mainnetHelpers'
+import { createPythOracle, PYTH_ETH_USD_PRICE_FEED } from '../../helpers/oracleHelpers'
+import { deployPythOracleFactory } from '../../helpers/setupHelpers'
 
 const ORACLE_STARTING_TIMESTAMP = BigNumber.from(1646456563)
 
@@ -105,11 +104,6 @@ async function getKeeperOracle(): Promise<[PythFactory, KeeperOracle]> {
   vars.oracle = oracle_
   return [pythOracleFactory, keeperOracle]
 }
-
-/*async function resetSubOracles(ethSubOracle: FakeContract<IOracleProvider>, btcSubOracle: FakeContract<IOracleProvider>): Promise<void> {
-  resetEthSubOracle(ethSubOracle, INITIAL_ORACLE_VERSION_ETH)
-  resetBtcSubOracle(btcSubOracle, INITIAL_ORACLE_VERSION_BTC)
-}*/
 
 if (process.env.FORK_NETWORK === undefined) {
   RunInvokerTests(
