+++ packages/periphery/test/integration/TriggerOrders/Manager.test.ts
@@ -4,16 +4,16 @@ import { anyValue } from '@nomicfoundation/hardhat-chai-matchers/withArgs'
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers'
 import HRE from 'hardhat'
 
-import { advanceBlock, currentBlockTimestamp, increase } from '../../../../../common/testutil/time'
-import { getEventArguments, getTimestamp } from '../../../../../common/testutil/transaction'
-import { parse6decimal } from '../../../../../common/testutil/types'
+import { advanceBlock, currentBlockTimestamp, increase } from '../../../../common/testutil/time'
+import { getEventArguments, getTimestamp } from '../../../../common/testutil/transaction'
+import { parse6decimal } from '../../../../common/testutil/types'
 
 import { IERC20Metadata, IMarketFactory, IMarket, IOracleProvider } from '@perennial/v2-core/types/generated'
 import { IKeeperOracle } from '@perennial/v2-oracle/types/generated'
-import { IEmptySetReserve, IManager, IOrderVerifier } from '../../../../types/generated'
-import { PlaceOrderActionStruct } from '../../../../types/generated/contracts/TriggerOrders/Manager'
+import { IEmptySetReserve, IManager, IOrderVerifier } from '../../../types/generated'
+import { PlaceOrderActionStruct } from '../../../types/generated/contracts/TriggerOrders/Manager'
 
-import { signAction, signCancelOrderAction, signPlaceOrderAction } from '../../../helpers/TriggerOrders/eip712'
+import { signAction, signCancelOrderAction, signPlaceOrderAction } from '../../helpers/TriggerOrders/eip712'
 import {
   Compare,
   compareOrders,
@@ -21,10 +21,10 @@ import {
   MAGIC_VALUE_CLOSE_POSITION,
   orderFromStructOutput,
   Side,
-} from '../../../helpers/TriggerOrders/order'
-import { advanceToPrice } from '../../../helpers/oracleHelpers'
+} from '../../helpers/TriggerOrders/order'
+import { advanceToPrice } from '../../helpers/oracleHelpers'
 import { Address } from 'hardhat-deploy/dist/types'
-import { impersonate } from '../../../../../common/testutil'
+import { impersonate } from '../../../../common/testutil'
 import { FixtureVars } from './setupTypes'
 
 const MAX_FEE = utils.parseEther('0.88')
